requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
anthropic==0.40.0
openai==1.54.0
youtube-transcript-api==0.6.3
yt-dlp==2024.10.22
httpx==0.27.2
aiofiles==24.1.0
psycopg2-binary==2.9.9
